+++ packages/core/src/request-builder/header-builder/index.ts
@@ -1,7 +1,7 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
 
 export * from './authorization-header';
-export * from './auth-headers';
 export * from './csrf-token-header';
 export * from './csrf-headers';
 export * from './header-builder';
+export * from './headers-util';
